Move functions from webservice_default.go to separate files
The webservice_default.go file (and the associated tests file) grew too large.
Adding the next function would bring it over 1000 lines already.
Split functions (and tests) into separate logically grouped files.

Client code should be unaffected by this,
packages and import paths remain the same.
