+++ infocmdb/response.go
@@ -1,0 +1,5 @@
+package infocmdb
+
+type responseId struct {
+	Id int `json:"id,string"`
+}
